+++ TaskManagmentSystemApiProject/TaskManagmentSystemApiProject/Controllers/TaskController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using TaskManagmentSystemApiProject.Dto;
 using TaskManagmentSystemApiProject.Interfaces;
@@ -16,15 +17,18 @@ namespace TaskManagmentSystemApiProject.Controllers
         private readonly ITaskRepository _taskrepos;
         private readonly IMapper _mapper;
         private readonly IUserService _service;
+        private readonly IUserRepository _userrepos;
 
-        public TaskController(ITaskRepository taskrepos, IMapper mapper, IUserService service)
+        public TaskController(ITaskRepository taskrepos, IUserRepository userrepos, IMapper mapper, IUserService service)
         {
             _taskrepos = taskrepos;
             _mapper = mapper;
             _service = service;
+            _userrepos = userrepos;
         }
         // GET: api/tasks
         [HttpGet]
+        [Authorize]
         public IActionResult Get([FromQuery] Status? status, [FromQuery] Priority? priority, [FromQuery] int? assignedTo)
         {
             return Ok(_mapper.Map<List<TaskDto>>(_taskrepos.GetAllTasks(status, priority, assignedTo)));
@@ -32,6 +36,7 @@ namespace TaskManagmentSystemApiProject.Controllers
 
         // GET api/tasks/5
         [HttpGet("{id}")]
+        [Authorize]
         public IActionResult Get(int id)
         {
             return Ok(_mapper.Map<TaskDto>(_taskrepos.GetTaskById(id)));
@@ -39,23 +44,48 @@ namespace TaskManagmentSystemApiProject.Controllers
 
         // POST api/tasks
         [HttpPost]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Post([FromBody] TaskDto task)
         {
+            if (task == null)
+            {
+                ModelState.AddModelError("", "Data has not been sent");
+                return BadRequest(ModelState);
+            }
+            task.CreatedById = _userrepos.GetUserByEmail(_service.GetEmail()).Id;
             return Ok(_taskrepos.CreateTask(_mapper.Map<Models.Task>(task)));
         }
 
         // PUT api/tasks/5
         [HttpPut("{id}")]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Put(int id, [FromBody] TaskDto task)
         {
-            return Ok(_taskrepos.UpdateTask(id, _mapper.Map<Models.Task>(task)));
+            if (task == null)
+            {
+                ModelState.AddModelError("", "Data has not been sent");
+                return BadRequest(ModelState);
+            }
+            if (_userrepos.GetUserByEmail(_service.GetEmail()).Id == _taskrepos.GetTaskById(id).CreatedBy.Id || _service.GetRole() == "Admin")
+            {
+                return Ok(_taskrepos.UpdateTask(id, _mapper.Map<Models.Task>(task)));
+            }
+            ModelState.AddModelError("", "You don't have permission to do that");
+            return StatusCode(403, ModelState);
         }
 
         // DELETE api/tasks/5
         [HttpDelete("{id}")]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Delete(int id)
         {
-            return Ok(_taskrepos.DeleteTask(id));
+            if (_userrepos.GetUserByEmail(_service.GetEmail()).Id == _taskrepos.GetTaskById(id).CreatedBy.Id || _service.GetRole() == "Admin")
+            {
+                return Ok(_taskrepos.DeleteTask(id));
+            }
+            ModelState.AddModelError("", "You don't have permission to do that");
+            return StatusCode(403, ModelState);
+
         }
     }
 }
